requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.15
psycopg[binary]>=3.1
python-dotenv>=1.0
celery>=5.3
redis>=5.0
requests>=2.32
django-auth-adfs>=1.16
